pyproject.toml
[tool.poetry]
name = "dnd_character_creator"
version = "0.1.0"
description = ""
authors = ["Tesla2000 <fratajczak124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.7.1"
dynamic-executor = "^1.2.6"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
beautifulsoup4 = "^4.12.3"
more-itertools = "^10.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dnd_character_creator = "dnd_character_creator:main"
